defaultConfig 내 버전 관련 설정들을 SemanticVersioningConventionPlugin으로 옮기기
