+++ build-logic/convention/src/main/kotlin/SemanticVersioningConventionPlugin.kt
@@ -1,10 +1,28 @@
+import com.android.build.gradle.AppExtension
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.gradle.kotlin.dsl.configure
+import org.gradle.kotlin.dsl.getByType
+import java.nio.file.Files
+import java.nio.file.Paths
+import java.util.Properties
 
 class SemanticVersioningConventionPlugin: Plugin<Project> {
     override fun apply(target: Project) {
         with(target) {
             extensions.create("versioning", SemanticVersioningUtils::class.java)
+            extensions.configure<AppExtension> {
+                val versionProps = Properties().apply {
+                    load(Files.newBufferedReader(Paths.get(rootProject.rootDir.toString(), "version.properties")))
+                }
+
+                defaultConfig {
+                    val propertyVersionName = versionProps.getProperty("snuttVersionName")
+                    versionName = propertyVersionName
+                    versionCode = extensions.getByType<SemanticVersioningUtils>()
+                        .semanticVersionToSerializedCode(propertyVersionName)
+                }
+            }
         }
     }
 }
